Module Tezos_dal_node_services.Operator_profile

type t

An operator DAL node can play three different roles:

attester for some pkh: checks that the shards assigned to this pkh are published

slot producer for some slot index: splits slots into shards and publishes the shards

slot observer for some slot index: collects the shards corresponding to some slot index, reconstructs slots when enough shards are seen, and republishes missing shards.

A single DAL node can play several of these roles at once. We call a profile the set of roles played by a DAL node and represent it as a triple of sets.

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val empty : t

The empty operator; an operator with this profile does nothing

val is_empty : t -> bool

val has_attester : t -> bool

has_attester operator returns true if there is a public key in the attester field of operator

val has_producer : t -> bool

has_producer operator returns true if there is a slot index in the producer field of operator

val has_observer : t -> bool

has_observer operator returns true if there is a slot index in the observer field of operator

val attester_only : t -> bool

attester_only operator returns true if operator has an attester role, and no producer, not observer roles.

val attesters : t -> Tezos_base.TzPervasives.Signature.Public_key_hash.Set.t

attesters operator returns the set of attesters registered within the attester profile, if any.

val producer_slot_out_of_bounds : int -> t -> int option

producer_slot_out_of_bounds n op returns the first slot index that for producer that is not between 0 and n - 1; it returns None if no slot index was outside these bounds

val is_observed_slot : int -> t -> bool

is_observed_slot i op returns true if and only if op contains an observer for slot index i

val can_publish_on_slot_index : int -> t -> bool

can_publish_on_slot_index slot_index op returns true if and only if op contains an observer or a producet for slot index slot_index

val get_all_slot_indexes : t -> int list

Returns all slot indexes used in the profile, whether they are for producer or observer

val make : ?attesters:Tezos_base.TzPervasives.Signature.Public_key_hash.t list -> ?producers:int list -> ?observers:int list -> unit -> t

make ~attesters ~producers ~observers () returns an operator profile for a node that is an attester for the public keys in attesters, a producer for each slot indexes in producers, and an observer for all slot indexes in observer. When a list is empty or not specified, no value for the corresponding role will be included in the result ; make () returns empty

val merge : ?on_new_attester: (Tezos_base.TzPervasives.Signature.Public_key_hash.t -> unit) -> t -> t -> t

merge ~on_new_attester op1 op2 returns an operator profile that contains both op1 & op2. on_new_attester is a function triggered for each public key in op2 that is not already in op1