Module Tezos_dal_node_services.Types

module Cryptobox = Tezos_crypto_dal.Cryptobox

type level = int32

A Tezos level.

type slot_index = int

An index of a DAL slot header.

type page_index = int

An index of a DAL page.

module Slot_id : sig ... end

An ID associated to a slot or to its commitment.

type slot_id = Slot_id.t

module Topic : sig ... end

A topic is defined by a public key hash of an attester and a slot index.

module Message_id : sig ... end

A message id uniquely identifies a share whose commitment is included in an L1 block. It is defined by a tuple containing the commitment, the level at which the commitment is successfully included in an L1 block, the corresponding slot index, the shard index, as well as the public key hash pkh of the delegate expected to attest it.

module Message : sig ... end

A message is a portion of an encoded slot's data. It's basically a shard without the corresponding index. The proof that the corresponding shard belong to the commitment (part of the message id) is also part of the message.

module Peer : sig ... end

A peer from the point of view of gossipsub.

module Point : sig ... end

A point is made of an IP address and a port. Only the worker knows about the notion. The automaton only sees peers (i.e. cryptographic identities of nodes).

module Span : sig ... end

module Time : sig ... end

module Score : sig ... end

type slot_set = {

slots : bool list;

published_level : int32;

}

A set of slots, represented by a list of booleans (false for not in the set). It is used for instance to record which slots are deemed available by an attester. The level at which the slots have been published is also given.

type attestable_slots =

| Attestable_slots of slot_set

| Not_in_committee

The set of attestable slots of an attester (which may not necessarily be in the committee for a given level).

type shard_index = int

An index of a DAL shard

type header_status = [

| `Waiting_attestation
(*
The slot header was included and applied in a finalized L1 block but remains to be attested.
*)

| `Attested
(*
The slot header was included in an L1 block and attested.
*)

| `Unattested
(*
The slot header was included in an L1 block but not timely attested.
*)

]

The status of a header a DAL node is aware of:

type profile =

| Bootstrap
(*
The bootstrap profile facilitates peer discovery in the DAL network. Note that bootstrap nodes are incompatible with attester/producer/observer profiles as bootstrap nodes are expected to connect to all the meshes with degree 0.
*)

| Operator of Operator_profile.t

| Random_observer

DAL node can track one or many profiles that correspond to various modes that the DAL node would operate in.

type slot_header = {

slot_id : slot_id;

commitment : Cryptobox.Commitment.t;

status : header_status;

}

Information associated to a slot header in the RPC services of the DAL node.

type with_proof = {

with_proof : bool;

}

The with_proof flag is associated to shards computation. It indicates whether we also compute shards' proofs or not.

val slot_id_query : (level option * shard_index option) Resto.Query.t

val slot_query : < padding : char ; slot_index : slot_index option > Resto.Query.t

val wait_query : < wait : bool > Resto.Query.t

val connected_query : < connected : bool > Resto.Query.t

val all_query : < all : bool > Resto.Query.t

val opt_header_status_query : header_status option Resto.Query.t

val slot_encoding : Cryptobox.slot Tezos_base.TzPervasives.Data_encoding.t

val slot_header_encoding : slot_header Tezos_base.TzPervasives.Data_encoding.t

val slot_id_encoding : slot_id Tezos_base.TzPervasives.Data_encoding.t

val header_status_encoding : header_status Tezos_base.TzPervasives.Data_encoding.t

val profile_encoding : profile Tezos_base.TzPervasives.Data_encoding.t

val with_proof_encoding : with_proof Tezos_base.TzPervasives.Data_encoding.t

val attestable_slots_encoding : attestable_slots Tezos_base.TzPervasives.Data_encoding.t

val pp_header_status : Stdlib.Format.formatter -> header_status -> unit

module Store : sig ... end

module P2P : sig ... end

module Gossipsub : sig ... end

module Version : sig ... end

module Health : sig ... end