Module Types.Message

A message is a portion of an encoded slot's data. It's basically a shard without the corresponding index. The proof that the corresponding shard belong to the commitment (part of the message id) is also part of the message.

type t = {

share : Cryptobox.share;

shard_proof : Cryptobox.shard_proof;

}

include Tezos_base.TzPervasives.PRINTABLE with type t := t

val pp : Stdlib.Format.formatter -> t -> unit

include Tezos_base.TzPervasives.ENCODABLE with type t := t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t