Module Types.Message_id

A message id uniquely identifies a share whose commitment is included in an L1 block. It is defined by a tuple containing the commitment, the level at which the commitment is successfully included in an L1 block, the corresponding slot index, the shard index, as well as the public key hash pkh of the delegate expected to attest it.

Note that pkh is used to be able to directly infer a topic from a message id. It could be retrieved from L1 using the level. But, we decide to provide it directly in this first version.

type t = {

commitment : Cryptobox.Commitment.t;

level : int32;

slot_index : int;

shard_index : int;

pkh : Tezos_base.TzPervasives.Signature.Public_key_hash.t;

}

include Tezos_base.TzPervasives.PRINTABLE with type t := t

val pp : Stdlib.Format.formatter -> t -> unit

include Tezos_base.TzPervasives.ENCODABLE with type t := t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

include Tezos_base.TzPervasives.COMPARABLE with type t := t

val (=) : t -> t -> bool

x = y iff compare x y = 0

val (<>) : t -> t -> bool

x <> y iff compare x y <> 0

val (<) : t -> t -> bool

x < y iff compare x y < 0

val (<=) : t -> t -> bool

x <= y iff compare x y <= 0

val (>=) : t -> t -> bool

x >= y iff compare x y >= 0

val (>) : t -> t -> bool

x > y iff compare x y > 0

val compare : t -> t -> int

compare an alias for the functor parameter's compare function

val equal : t -> t -> bool

equal x y iff compare x y = 0

val max : t -> t -> t

max x y is x if x >= y otherwise it is y

val min : t -> t -> t

min x y is x if x <= y otherwise it is y

module Set : Tezos_base.TzPervasives.Set.S with type elt = t

module Map : Tezos_base.TzPervasives.Map.S with type key = t

val get_topic : t -> Topic.t