Module Metadata.Connection

type t = {

advertised_net_addr : Tezos_base.P2p_addr.t option;
(*
The public address for which the local node can be reached from the outside. This is useful if the node is behind a NAT or a load balancer for example.
*)

advertised_net_port : int option;
(*
The port at which the local node can be reached. It is a bit redundant since the handshaking protocol already exchange this piece of information.
*)

is_bootstrap_peer : bool;
(*
true if the node advertises itself as a bootstrap node. This is to prevent a race condition from Gossipsub where a node may send full messages to a bootstrap node while this is not necessary.
*)

}

connection_metadata type.

A value of this type is exchanged through the handshake protocol of the P2P.

The advertised_net_port is not mandatory, as it is already sent via the first P2P message after a connection is authenticated. But, we decide to duplicate the information here for consistency. The is_bootstrap_peer indicates whether the remote peer has a bootstrap profile or not.

include Tezos_base.TzPervasives.ENCODABLE with type t := t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val config : t -> t Tezos_base.P2p_params.conn_meta_config