Module Types.Peer

A peer from the point of view of gossipsub.

type t = {

peer_id : Tezos_base.P2p_peer.Id.t;

maybe_reachable_point : Tezos_base.P2p_point.Id.t;

}

For incoming connections, we know the peer is reachable, and the peer's address and the port are provided by the octez-p2p layer.

For outgoing connections, if the remote peers has not specified its address, nor its port, then the peer's address is <remote_socket_addr>:<default_dal_p2p_port> (at the time of writing 11732). However, the remote node's user can override both the address and the port if desired and in that case, the specified values will be used.

Comparison is not a structural one, instead only the peer_id is used.

include Tezos_base.TzPervasives.PRINTABLE with type t := t

val pp : Stdlib.Format.formatter -> t -> unit

include Tezos_base.TzPervasives.ENCODABLE with type t := t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

include Tezos_base.TzPervasives.COMPARABLE with type t := t

val (=) : t -> t -> bool

x = y iff compare x y = 0

val (<>) : t -> t -> bool

x <> y iff compare x y <> 0

val (<) : t -> t -> bool

x < y iff compare x y < 0

val (<=) : t -> t -> bool

x <= y iff compare x y <= 0

val (>=) : t -> t -> bool

x >= y iff compare x y >= 0

val (>) : t -> t -> bool

x > y iff compare x y > 0

val compare : t -> t -> int

compare an alias for the functor parameter's compare function

val equal : t -> t -> bool

equal x y iff compare x y = 0

val max : t -> t -> t

max x y is x if x >= y otherwise it is y

val min : t -> t -> t

min x y is x if x <= y otherwise it is y

module Set : Tezos_base.TzPervasives.Set.S with type elt = t

module Map : Tezos_base.TzPervasives.Map.S with type key = t